# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vital-AGRO Weighings API",
    "version": "1.0.1",
    "description": "This is a description of Vital-AGRO's weghings API server, based on the OpenAPI 3.0 specification.\n\nSome useful links:\n- [repository](https://github.com/AegisVP/weighings-VA)"
  },
  "servers": [
    {
      "url": "https://vitalagro.onrender.com/api"
    }
  ],
...
  "paths": {
...
    "/user/login": {
      "post": {
        "tags": ["user"],
        "summary": "Logs user into the system",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "example@email.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "a1b2c3"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userAuthInfo"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email/password supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseMessage"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/user/logout": {
      "post": {
        "tags": ["user"],
        "summary": "Logs out current logged in user session",
        "operationId": "logoutUser",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "\\_token\\_ is the JWT token issued to the current user during login",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer _token_"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAuth"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/user": {
      "patch": {
        "tags": ["user"],
        "summary": "Update user subscription level",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "\\_token\\_ is the JWT token issued to the current user during login",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Bearer _token_"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "example@email.com"
                  },
                  "subscription": {
                    "type": "string",
                    "example": "a01b02c03"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "$ref": "#/components/responses/InvalidAuth"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "InvalidAuth": {
        "description": "Invalid Authorization",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Not authorized"
                }
              }
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ServerErrorMessage"
            }
          }
        }
      }
    },
    "schemas": {
...
      "userAuthInfo": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "aa11bb22cc33"
          }
        }
      },
      "ResponseMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Operation status message"
          }
        }
      },
      "ServerErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Internal server error"
          },
          "details": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Error title"
              },
              "message": {
                "type": "string",
                "example": "Error description"
              }
            }
          }
        }
      }
    }
  }
}